Convertendo CountDown para context
