+++ src/page/Home/index.tsx
@@ -1,33 +1,29 @@
-import { zodResolver } from '@hookform/resolvers/zod'
-import { useForm } from 'react-hook-form'
+// import { zodResolver } from '@hookform/resolvers/zod'
+// import { useForm } from 'react-hook-form'
 import { HandPalm, Play } from 'phosphor-react'
-import { useEffect, useState } from 'react'
-import * as zod from 'zod'
-import { differenceInSeconds } from 'date-fns'
+import { createContext, useState } from 'react'
+// import * as zod from 'zod'
 
 import {
-  FormContainer,
+  // FormContainer,
   HomeContainer,
-  CountdownContainer,
-  Separetor,
   StartCountDownButton,
-  TaskInput,
-  MinutAmountInput,
   StoptCountDownButton,
 } from './styles'
+import { CountDown } from './components/CountDown'
 
-const schema = zod.object({
-  minutAmount: zod
-    .number()
-    .min(1, 'Ciclo deve ser minimo de 1 minutos')
-    .max(60, 'Ciclo deve ser maximo 60 minutos'),
-  task: zod
-    .string()
-    .min(2, 'Titulo deve ser de no minimut 2 caracteres')
-    .max(10, 'Titulo deve ser de no máximo 10 caracteres'),
-})
+// const schema = zod.object({
+//   minutAmount: zod
+//     .number()
+//     .min(1, 'Ciclo deve ser minimo de 1 minutos')
+//     .max(60, 'Ciclo deve ser maximo 60 minutos'),
+//   task: zod
+//     .string()
+//     .min(2, 'Titulo deve ser de no minimut 2 caracteres')
+//     .max(10, 'Titulo deve ser de no máximo 10 caracteres'),
+// })
 
-type NewCicleFormData = zod.infer<typeof schema>
+// type NewCicleFormData = zod.infer<typeof schema>
 
 interface Cycle {
   id: string
@@ -37,52 +33,59 @@ interface Cycle {
   interrupDate: Date
   finishedDate: Date
 }
+interface CycleContextType {
+  activeCycle: Cycle | undefined
+  activeCycleId: string | null
+  markCurrentCycleAsFinished: () => void
+}
+export const CycleContext = createContext({} as CycleContextType)
 
 export function Home() {
-  const {
-    register,
-    handleSubmit,
-    watch,
-    formState: { errors },
-    reset,
-  } = useForm<NewCicleFormData>({
-    resolver: zodResolver(schema),
-    defaultValues: {
-      task: '',
-      minutAmount: 0,
-    },
-  })
+  // const {
+  //   // register,
+  //   // handleSubmit,
+  //   // watch,
+  //   // reset,
+  // } = useForm<NewCicleFormData>({
+  //   resolver: zodResolver(schema),
+  //   defaultValues: {
+  //     task: '',
+  //     minutAmount: 0,
+  //   },
+  // })
 
   const [cycle, setCycle] = useState<Cycle[]>([])
   const [activeCycleId, setActiveCycleId] = useState<string | null>(null)
   const activeCycle = cycle.find((cycle) => cycle.id === activeCycleId)
-  const [amountSecondsPassed, setAmountsecondsPassed] = useState(0)
 
-  const totalSeconds = activeCycle ? activeCycle.minutAmount * 60 : 0
-  const currentSeconds = activeCycle ? totalSeconds - amountSecondsPassed : 0
+  // const task = watch('task')
+  // const isDisableTask = !task
 
-  const amountMinutes = Math.floor(currentSeconds / 60)
-  const amountSeconds = currentSeconds % 60
+  // const handleCreatedNewCicle = (data: NewCicleFormData) => {
+  //   const id = String(new Date().getTime())
 
-  const minutes = String(amountMinutes).padStart(2, '0')
-  const seconds = String(amountSeconds).padStart(2, '0')
+  //   const newCycle: Cycle = {
+  //     id,
+  //     task: data.task,
+  //     minutAmount: data.minutAmount,
+  //     date: new Date(),
+  //   }
+  //   setCycle((state) => [...state, newCycle])
+  //   setActiveCycleId(id)
+  //   setAmountsecondsPassed(0)
+  //   reset()
+  // }
 
-  const task = watch('task')
-  const isDisableTask = !task
-
-  const handleCreatedNewCicle = (data: NewCicleFormData) => {
-    const id = String(new Date().getTime())
-
-    const newCycle: Cycle = {
-      id,
-      task: data.task,
-      minutAmount: data.minutAmount,
-      date: new Date(),
-    }
-    setCycle((state) => [...state, newCycle])
-    setActiveCycleId(id)
-    setAmountsecondsPassed(0)
-    reset()
+  function markCurrentCycleAsFinished() {
+    setCycle((status) =>
+      status.map((cycle) => {
+        if (cycle.id === activeCycleId) {
+          return { ...cycle, finishedDate: new Date() }
+        } else {
+          return cycle
+        }
+      }),
+    )
   }
 
   function handleInterruptedcycles() {
@@ -98,47 +101,10 @@ export function Home() {
     setActiveCycleId(null)
   }
 
-  useEffect(() => {
-    let interval: number
-    if (activeCycle) {
-      interval = setInterval(() => {
-        const secondsDiferents = differenceInSeconds(
-          new Date(),
-          activeCycle.date,
-        )
-
-        if (secondsDiferents >= totalSeconds) {
-          setCycle((status) =>
-            status.map((cycle) => {
-              if (cycle.id === activeCycleId) {
-                return { ...cycle, finishedDate: new Date() }
-              } else {
-                return cycle
-              }
-            }),
-          )
-          setAmountsecondsPassed(totalSeconds)
-          clearInterval(interval)
-        } else {
-          setAmountsecondsPassed(secondsDiferents)
-        }
-      }, 1000)
-    }
-    return () => {
-      clearInterval(interval)
-    }
-  }, [activeCycle, totalSeconds, activeCycleId])
-
-  useEffect(() => {
-    if (activeCycle) {
-      document.title = `${minutes}: ${seconds}`
-    }
-  }, [minutes, seconds, activeCycle])
-
   return (
     <HomeContainer>
-      <form onSubmit={handleSubmit(handleCreatedNewCicle)}>
-        <FormContainer>
+      <form /* onSubmit={handleSubmit(handleCreatedNewCicle)} */>
+        {/* <FormContainer>
           <label htmlFor="task">Vou trabalhar em</label>
           <TaskInput
             type="text"
@@ -166,23 +132,19 @@ export function Home() {
           {errors.minutAmount?.message && <p>{errors.minutAmount?.message}</p>}
 
           <span>minutos.</span>
-        </FormContainer>
-
-        <CountdownContainer>
-          <span>{minutes[0]}</span>
-          <span>{minutes[1]}</span>
-          <Separetor>:</Separetor>
-          <span>{seconds[0]}</span>
-          <span>{seconds[1]}</span>
-        </CountdownContainer>
-
+        </FormContainer> */}
+        <CycleContext.Provider
+          value={{ activeCycle, activeCycleId, markCurrentCycleAsFinished }}
+        >
+          <CountDown />
+        </CycleContext.Provider>
         {activeCycle ? (
           <StoptCountDownButton onClick={handleInterruptedcycles} type="button">
             <HandPalm size={24} />
             Interromper
           </StoptCountDownButton>
         ) : (
-          <StartCountDownButton disabled={isDisableTask} type="submit">
+          <StartCountDownButton /* disabled={isDisableTask} */ type="submit">
             <Play size={24} />
             Começar
           </StartCountDownButton>
