+++ src/page/Home/styles.ts
@@ -54,33 +54,6 @@ export const TaskInput = styled(BaseInput)`
 export const MinutAmountInput = styled(BaseInput)`
   width: 4rem;
 `
-
-export const CountdownContainer = styled.div`
-  font-family: 'Roboto Mono', monospace;
-  font-size: 10rem;
-  line-height: 8rem;
-  color: ${(props) => props.theme['gray-100']};
-
-  display: flex;
-  gap: 1rem;
-
-  span {
-    background: ${(props) => props.theme['gray-700']};
-    padding: 2rem 1rem;
-    border-radius: 8px;
-  }
-`
-
-export const Separetor = styled.div`
-  padding: 2rem 0;
-  color: ${(props) => props.theme['green-700']};
-
-  width: 4rem;
-  overflow: hidden;
-  display: flex;
-  justify-content: center;
-`
-
 export const BasicCountDownButton = styled.button`
   width: 100%;
   border: 0;
